Renames reducers file
