+++ src/shared/Entities/reducers.js
